pyproject.toml
[project]
name = "ashwinsundar"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-frontmatter>=1.1",
    "markdown-it-py>=3.0",
    "mdit-py-plugins>=0.4",
    "toml>=0.10",
]

